Open feedback form upon tapping "Send Feedback" in the About page
