+++ DuckDuckGo/Common/Extensions/URLExtension.swift
@@ -361,20 +361,6 @@ extension URL {
         return try? getParameter(name: DuckDuckGoParameters.search.rawValue)
     }
 
-    // MARK: - Feedback
-
-#if FEEDBACK
-
-    static var feedback: URL {
-    #if BETA
-        return URL(string: "https://www.surveymonkey.com/r/WTBLLJR")!
-    #else
-        return URL(string: "https://form.asana.com/?k=HzdxNoIgDZUBf4w0_cafIQ&d=137249556945")!
-    #endif
-    }
-
-#endif
-
     // MARK: - Punycode
 
     var punycodeDecodedString: String? {
